Add `include_snippet` and `flexible_include` plugins.
`include_snippet` allows part of a file, i.e. a snippet from a file, to be added to a document, it inherits from `flexible_include`

`flexible_include` provides access to files that are outside the `_includes` directory.  Key if we're to include code snippets from real repo code.

`flexible_include` has had some functionality disabled, as they are not needed:
 - loading absolute paths
 - loading from home dir
 - executing commands.